pyproject.toml
[project]
name = "visflow"
version = "1.0.0"
description = ""
authors = [
    {name = "BoChen SHEN",email = "6goddddddd@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<=3.12"
dependencies = [
    "pydantic (>=2.11.7,<3.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "seaborn (>=0.13.2,<0.14.0)",
    "onnxruntime (>=1.22.1,<2.0.0)",
    "onnx (>=1.19.0,<2.0.0)",
    "scikit-learn (>=1.7.1,<2.0.0)",
    "matplotlib (>=3.10.6,<4.0.0)",
    "numpy (>=2.3.2,<3.0.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "grad-cam (>=1.5.5,<2.0.0)",
    "torch (>=2.8.0,<3.0.0)",
    "torchvision (>=0.23.0,<0.24.0)",
    "torchsummary (>=1.5.1,<2.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
